+++ Src/ImageQT/DllInterop/Mac/ObjectCRuntime.cs
@@ -21,7 +21,13 @@ internal partial class ObjectCRuntime
     public static partial void ObjCMsgSend(IntPtr receiver, IntPtr selector, IntPtr arg1);
 
     [LibraryImport(_dllName, EntryPoint = "objc_msgSend")]
+    public static partial void ObjCMsgSend(IntPtr receiver, IntPtr selector, NSImageView arg1);
+
+    [LibraryImport(_dllName, EntryPoint = "objc_msgSend")]
     public static partial void ObjCMsgSend(NSWindow receiver, IntPtr selector, IntPtr arg1);
+
+    [LibraryImport(_dllName, EntryPoint = "objc_msgSend")]
+    public static partial void ObjCMsgSend(NSImageView receiver, IntPtr selector, IntPtr arg1);
 
     [LibraryImport(_dllName, EntryPoint = "objc_msgSend")]
     public static partial void ObjCMsgSend(NSWindow receiver, IntPtr selector, NSCustomClass arg1);
